refactor: move x,y domains to types/xy_chart/domains
