+++ src/state/chart_state.ts
@@ -8,8 +8,8 @@ import {
   mergeYCustomDomainsByGroupId,
 } from '../types/xy_chart/utils/axis_utils';
 import { CanvasTextBBoxCalculator } from '../utils/bbox/canvas_text_bbox_calculator';
-import { XDomain } from '../lib/series/domains/x_domain';
-import { YDomain } from '../lib/series/domains/y_domain';
+import { XDomain } from '../types/xy_chart/domains/x_domain';
+import { YDomain } from '../types/xy_chart/domains/y_domain';
 import { computeLegend, LegendItem } from '../types/xy_chart/legend/legend';
 import {
   AreaGeometry,
