+++ src/types/xy_chart/domains/y_domain.ts
@@ -1,9 +1,9 @@
-import { BasicSeriesSpec, DomainRange } from '../../../types/xy_chart/utils/specs';
+import { BasicSeriesSpec, DomainRange } from '../utils/specs';
 import { GroupId, SpecId } from '../../../utils/ids';
 import { ScaleContinuousType, ScaleType } from '../../../utils/scales/scales';
-import { isCompleteBound, isLowerBound, isUpperBound } from '../../../types/xy_chart/utils/axis_utils';
+import { isCompleteBound, isLowerBound, isUpperBound } from '../utils/axis_utils';
 import { BaseDomain } from './domain';
-import { RawDataSeries } from '../series';
+import { RawDataSeries } from '../../../lib/series/series';
 import { computeContinuousDataDomain } from '../../../utils/domain';
 import { identity } from '../../../utils/commons';
 import { sum } from 'd3-array';
