+++ src/types/xy_chart/utils/axis_utils.ts
@@ -1,5 +1,5 @@
-import { XDomain } from '../../../lib/series/domains/x_domain';
-import { YDomain } from '../../../lib/series/domains/y_domain';
+import { XDomain } from '../domains/x_domain';
+import { YDomain } from '../domains/y_domain';
 import { computeXScale, computeYScales } from './scales';
 import {
   AxisSpec,
