+++ src/types/xy_chart/utils/scales.test.ts
@@ -1,6 +1,6 @@
 import { getGroupId } from '../../../utils/ids';
 import { ScaleType } from '../../../utils/scales/scales';
-import { XDomain } from '../../../lib/series/domains/x_domain';
+import { XDomain } from '../domains/x_domain';
 import { computeXScale, countBarsInCluster } from './scales';
 import { FormattedDataSeries } from '../../../lib/series/series';
