+++ src/utils/scales/scale_continuous.test.ts
@@ -1,4 +1,4 @@
-import { XDomain } from '../../lib/series/domains/x_domain';
+import { XDomain } from '../../types/xy_chart/domains/x_domain';
 import { computeXScale } from '../../types/xy_chart/utils/scales';
 import { Domain } from '../domain';
 import { DateTime, Settings } from 'luxon';
